fix(Onboarding): Cleaned onboarding creation process.
Removed unnecessary is_private property from POST /api/onboarding body and send is_internal in a way, when we check that the selected dform template is private and if it is not, then we send the client state which is controlled by checkbox. Also in case when the selected dform is private, the checkbox which control client state of internal state is disabled to prevent cases to create public onboarding from private dform template.